python/requirements.txt
Pillow~=7.2.0
pip~=20.1.1
setuptools~=49.2.0